Refactor code and begin bonus release
