+++ celebrate.rb
@@ -1,0 +1,5 @@
+module HighFive
+   def offer_high_five
+    "High five!"
+  end
+end
