+++ educator.rb
@@ -1,0 +1,18 @@
+class Educator
+  attr_reader :age, :salary, :phase, :target_raise
+  attr_accessor :name
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+end
